+++ Editor/src/App/Editor/UI/Tools/Keylogging.cpp
@@ -430,10 +430,10 @@ void KeyLogging::LoadKeylogs(const std::filesystem::path& path)
 		auto arr = iter->value.GetArray();
 		float t = arr[0].GetFloat();
 		MousePos mp;
-		mp.dx = arr[1].GetInt();
-		mp.dy = arr[2].GetInt();
-		mp.x = arr[3].GetInt();
-		mp.y = arr[4].GetInt();
+		mp.dx = static_cast<short>(arr[1].GetInt());
+		mp.dy = static_cast<short>(arr[2].GetInt());
+		mp.x = static_cast<short>(arr[3].GetInt());
+		mp.y = static_cast<short>(arr[4].GetInt());
 
 		m_mousePosition.push_back(std::pair(t, mp));
 	}
